Add data-depth as option, default 12
In commit b8efde0e7e74 (Fixes after testing with Basler Scout),
data-depth was set to mono8 as Scout support only 8bit depth.

Other basler cameras support mono12, and it is generally desirable to
utilize the full depth of the sensor, so set 12 as default and instead
add a switch to set it to a different depth when this is required.