Handle image-sequence
GigE-camera has support for capturing a sequence of images before stopping
the stream. This change enables basler-hr to do the same. Once a full
sequence has been captured, the camera needs to be stopped before it can
be started to capture a new sequence of images.

Once in img-series mode, the camera needs to have the series explicitly
disabled (-1) to move back to continous mode.